requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-ckeditor==5.6.1
django-elasticsearch-dsl==0.5.1
django-extensions==2.1.4
django-js-asset==1.1.0
django-tinymce==2.7.0
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
gunicorn==19.9.0
ipaddress==1.0.22
Pillow==5.3.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydotplus==2.0.2
pyparsing==2.3.0
python-dateutil==2.7.5
python-slugify==1.2.6
pytz==2018.6
six==1.12.0
slugify==0.0.1
Unidecode==1.0.23
urllib3==1.24.1
whitenoise==4.1.2
